fix(nvcc): append to `cpp_compile` `PATH`
Instead of reducing the PATH of the `cpp_compile` action to only
the dir of the host compiler, we can query we toolchain for the
configured PATH and append the directory of the host compiler.

This way 1) nvcc can still find the host compiler from the PATH
and 2) a potential compiler wrapper script that acts as the
host compiler can still find its dependencies.